React Maintenance: React Project Structure
