+++ The Road To React/hacker-stories/src/InputWithLabel.jsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+const InputWithLabel = ({ id, value, onInputChange, type = 'text', children }) => (
+  <>
+    <label htmlFor={id}>{children}</label>
+    <input id={id} type={type} value={value} onChange={onInputChange} />
+  </>
+);
+
+export { InputWithLabel };
