+++ The Road To React/hacker-stories/src/List.jsx
@@ -1,0 +1,34 @@
+import * as React from 'react';
+
+const List = ({ list, onRemoveItem }) =>
+(
+  <ul>
+    {list.map((item) =>
+      <Item
+        key={item.objectID}
+        item={item}
+        onRemoveItem={onRemoveItem}
+      />
+    )}
+  </ul>
+);
+
+const Item = ({ item, onRemoveItem }) => {
+  return (
+    <li >
+      <span>
+        <a href={item.url}>{item.title} </a>
+      </span>
+      <span>{item.author} </span>
+      <span>{item.num_comments} </span>
+      <span>{item.points} </span>
+      <span>
+        <button type='button' onClick={() => onRemoveItem(item)}>
+          Dismiss
+        </button>
+      </span>
+    </li>
+  );
+};
+
+export { List, Item };
