+++ The Road To React/hacker-stories/src/SearchForm.jsx
@@ -1,0 +1,24 @@
+import * as React from 'react';
+import { InputWithLabel } from './InputWithLabel';
+
+const SearchForm = ({
+  searchTerm, onSearchInput, onSearchSubmit
+}) => (
+  <form onSubmit={onSearchSubmit}>
+    <InputWithLabel
+      id="search"
+      label="Search"
+      value={searchTerm}
+      isFocused
+      onInputChange={onSearchInput}
+    >
+      <strong>Search: </strong>
+    </InputWithLabel>
+
+    <button type='submit' disabled={!searchTerm}>
+      Submit
+    </button>
+  </form>
+);
+
+export { SearchForm };
